+++ boofuzz/ifuzz_logger.py
@@ -173,3 +173,31 @@ class IFuzzLogger(object):
         :rtype: None
         """
         raise NotImplementedError
+
+    @abc.abstractmethod
+    def close_test_case(self):
+        """
+        Called after a test case has been completed. Can be used to inform the operator
+        or save the test case log.
+
+        :param None
+        :type None
+
+        :return: None
+        :rtype: None
+        """
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def close_test(self):
+        """
+        Called after a test has been completed. Can be used to inform the operator
+        or save the test log.
+
+        :param None
+        :type None
+
+        :return: None
+        :rtype: None
+        """
+        raise NotImplementedError