+++ boofuzz/sessions.py
@@ -1,8 +1,7 @@
-from __future__ import absolute_import
+from __future__ import absolute_import, print_function
 
 import cPickle
 import datetime
-import itertools
 import logging
 import os
 import re
@@ -10,6 +9,8 @@ import threading
 import time
 import traceback
 import zlib
+import socket
+import errno
 
 from tornado.httpserver import HTTPServer
 from tornado.ioloop import IOLoop
@@ -26,8 +27,6 @@ from . import pgraph
 from . import primitives
 from . import exception
 from .web.app import app
-
-DEFAULT_MAX_RECV = 8192
 
 
 class Target(object):
@@ -47,12 +46,13 @@ class Target(object):
         connection (itarget_connection.ITargetConnection): Connection to system under test.
     """
 
-    def __init__(self, connection, procmon=None, procmon_options=None, netmon=None):
+    def __init__(self, connection, procmon=None, procmon_options=None, netmon=None, max_recv_bytes=10000):
         self._fuzz_data_logger = None
 
         self._target_connection = connection
         self.procmon = procmon
         self.netmon = netmon
+        self.max_recv_bytes = max_recv_bytes
 
         # set these manually once target is instantiated.
         self.vmcontrol = None
@@ -110,7 +110,7 @@ class Target(object):
             for key in self.netmon_options.keys():
                 eval('self.netmon.set_%s(self.netmon_options["%s"])' % (key, key))
 
-    def recv(self, max_bytes=DEFAULT_MAX_RECV):
+    def recv(self, max_bytes=None):
         """
         Receive up to max_bytes data from the target.
 
@@ -120,6 +120,9 @@ class Target(object):
         Returns:
             Received data.
         """
+        if max_bytes is None:
+            max_bytes = self.max_recv_bytes
+
         if self._fuzz_data_logger is not None:
             self._fuzz_data_logger.log_info("Receiving...")
 
@@ -141,12 +144,12 @@ class Target(object):
             None
         """
         if self._fuzz_data_logger is not None:
-            self._fuzz_data_logger.log_send(data)
+            self._fuzz_data_logger.log_info("Sending {0} bytes...".format(len(data)))
 
         num_sent = self._target_connection.send(data=data)
 
         if self._fuzz_data_logger is not None:
-            self._fuzz_data_logger.log_info("{0} bytes sent".format(num_sent))
+            self._fuzz_data_logger.log_send(data[:num_sent])
 
     def set_fuzz_data_logger(self, fuzz_data_logger):
         """
@@ -277,9 +280,19 @@ class Session(pgraph.Graph):
         crash_threshold_request (int):  Maximum number of crashes allowed before a request is exhausted. Default 12.
         crash_threshold_element (int):  Maximum number of crashes allowed before an element is exhausted. Default 3.
         restart_sleep_time (int): Time in seconds to sleep when target can't be restarted. Default 5.
+        restart_callbacks (list of method): The registered method will be called after a failed post_test_case_callback
+                                           Default None.
+        pre_send_callbacks (list of method): The registered method will be called prior to each fuzz request.
+                                            Default None.
+        post_test_case_callbacks (list of method): The registered method will be called after each fuzz test case.
+                                                  Default None.
         web_port (int):         Port for monitoring fuzzing campaign via a web browser. Default 26000.
+        keep_web_open (bool):     Keep the webinterface open after session completion. Default True.
         fuzz_data_logger (fuzz_logger.FuzzLogger): DEPRECATED. Use fuzz_loggers instead.
         fuzz_loggers (list of ifuzz_logger.IFuzzLogger): For saving test data and results.. Default Log to STDOUT.
+        fuzz_db_keep_only_n_pass_cases (int): Minimize disk usage by only saving passing test cases
+                                              if they are in the n test cases preceding a failure or error.
+                                              Set to 0 to save after every test case (high disk I/O!). Default 0.
         receive_data_after_each_request (bool): If True, Session will attempt to receive a reply after transmitting
                                                 each non-fuzzed node. Default True.
         check_data_received_each_request (bool): If True, Session will verify that some data has
@@ -295,6 +308,8 @@ class Session(pgraph.Graph):
         ignore_connection_issues_when_sending_fuzz_data (bool): Ignore fuzz data transmission failures. Default True.
                                 This is usually a helpful setting to enable, as targets may drop connections once a
                                 message is clearly invalid.
+        reuse_target_connection (bool): If True, only use one target connection instead of reconnecting each test case.
+                                        Default False.
         target (Target):        Target for fuzz session. Target must be fully initialized. Default None.
 
         log_level (int):        DEPRECATED Unused. Logger settings are now configured in fuzz_data_logger.
@@ -308,11 +323,16 @@ class Session(pgraph.Graph):
     def __init__(self, session_filename=None, index_start=1, index_end=None, sleep_time=0.0,
                  restart_interval=0,
                  web_port=constants.DEFAULT_WEB_UI_PORT,
+                 keep_web_open=True,
                  crash_threshold_request=12,
                  crash_threshold_element=3,
                  restart_sleep_time=5,
+                 restart_callbacks=None,
+                 pre_send_callbacks=None,
+                 post_test_case_callbacks=None,
                  fuzz_data_logger=None,
                  fuzz_loggers=None,
+                 fuzz_db_keep_only_n_pass_cases=0,
                  receive_data_after_each_request=True,
                  check_data_received_each_request=False,
                  receive_data_after_fuzz=False,
@@ -320,11 +340,13 @@ class Session(pgraph.Graph):
                  ignore_connection_reset=False,
                  ignore_connection_aborted=False,
                  ignore_connection_issues_when_sending_fuzz_data=True,
+                 reuse_target_connection=False,
                  target=None,
                  ):
         self._ignore_connection_reset = ignore_connection_reset
         self._ignore_connection_aborted = ignore_connection_aborted
         self._ignore_connection_issues_when_sending_fuzz_data = ignore_connection_issues_when_sending_fuzz_data
+        self._reuse_target_connection = reuse_target_connection
         _ = log_level
         _ = logfile
         _ = logfile_level
@@ -337,6 +359,7 @@ class Session(pgraph.Graph):
         self.sleep_time = sleep_time
         self.restart_interval = restart_interval
         self.web_port = web_port
+        self._keep_web_open = keep_web_open
         self._crash_threshold_node = crash_threshold_request
         self._crash_threshold_element = crash_threshold_element
         self.restart_sleep_time = restart_sleep_time
@@ -348,7 +371,8 @@ class Session(pgraph.Graph):
         helpers.mkdir_safe(os.path.join(constants.RESULTS_DIR))
         self._run_id = datetime.datetime.utcnow().replace(microsecond=0).isoformat().replace(':', '-')
         self._db_filename = os.path.join(constants.RESULTS_DIR, 'run-{0}.db'.format(self._run_id))
-        self._db_logger = fuzz_logger_db.FuzzLoggerDb(self._db_filename)
+        self._db_logger = fuzz_logger_db.FuzzLoggerDb(db_filename=self._db_filename,
+                                                      num_log_cases=fuzz_db_keep_only_n_pass_cases)
 
         self._fuzz_data_logger = fuzz_logger.FuzzLogger(fuzz_loggers=[self._db_logger] + fuzz_loggers)
         self._check_data_received_each_request = check_data_received_each_request
@@ -356,9 +380,24 @@ class Session(pgraph.Graph):
         self._receive_data_after_fuzz = receive_data_after_fuzz
         self._skip_current_node_after_current_test_case = False
         self._skip_current_element_after_current_test_case = False
-        self._post_test_case_methods = []
 
-        self.web_interface_thread = self.build_webapp_thread(port=self.web_port)
+        if self.web_port is not None:
+            self.web_interface_thread = self.build_webapp_thread(port=self.web_port)
+
+        if pre_send_callbacks is None:
+            self._pre_send_methods = []
+        else:
+            self._pre_send_methods = pre_send_callbacks
+
+        if post_test_case_callbacks is None:
+            self._post_test_case_methods = []
+        else:
+            self._post_test_case_methods = post_test_case_callbacks
+
+        if restart_callbacks is None:
+            self._restart_methods = []
+        else:
+            self._restart_methods = restart_callbacks
 
         self.total_num_mutations = 0
         self.total_mutant_index = 0
@@ -623,9 +662,12 @@ class Session(pgraph.Graph):
         Returns:
             None
         """
-        self.server_init()
+        if self.web_port is not None:
+            self.server_init()
 
         try:
+            if self._reuse_target_connection:
+                self.targets[0].open()
             num_cases_actually_fuzzed = 0
             for fuzz_args in fuzz_case_iterator:
                 if self.total_mutant_index < self._index_start:
@@ -638,11 +680,19 @@ class Session(pgraph.Graph):
                         and self.restart_interval \
                         and num_cases_actually_fuzzed % self.restart_interval == 0:
                     self._fuzz_data_logger.open_test_step("restart interval of %d reached" % self.restart_interval)
-                    self.restart_target(self.targets[0])
+                    self._restart_target(self.targets[0])
 
                 self._fuzz_current_case(*fuzz_args)
 
                 num_cases_actually_fuzzed += 1
+            self._fuzz_data_logger.close_test()
+            if self._reuse_target_connection:
+                self.targets[0].close()
+
+            if self._keep_web_open and self.web_port is not None:
+                print("\nFuzzing session completed. Keeping webinterface up on localhost:{}".format(self.web_port),
+                      "\nPress ENTER to close webinterface")
+                raw_input()
         except KeyboardInterrupt:
             # TODO: should wait for the end of the ongoing test case, and stop gracefully netmon and procmon
             self.export_file()
@@ -825,7 +875,7 @@ class Session(pgraph.Graph):
                     self.total_mutant_index += skipped
                     self.fuzz_node.mutant_index += skipped
 
-            self.restart_target(target)
+            self._restart_target(target)
             return True
         else:
             return False
@@ -861,7 +911,6 @@ class Session(pgraph.Graph):
             session (Session): Session object calling post_send.
                 Useful properties include last_send and last_recv.
 
-            sock: DEPRECATED Included for backward-compatibility. Same as target.
             args: Implementations should include \*args and \**kwargs for forward-compatibility.
             kwargs: Implementations should include \*args and \**kwargs for forward-compatibility.
         """
@@ -869,10 +918,9 @@ class Session(pgraph.Graph):
         self._fuzz_data_logger.log_info("No post_send callback registered.")
 
     # noinspection PyMethodMayBeStatic
-    def pre_send(self, sock):
+    def _pre_send(self, target):
         """
-        Overload or replace this routine to specify actions to run prior to each fuzz request. The order of events is
-        as follows::
+        Execute custom methods to run prior to each fuzz request. The order of events is as follows::
 
             pre_send() - req - callback ... req - callback - post_send()
 
@@ -881,47 +929,71 @@ class Session(pgraph.Graph):
         @see: pre_send()
 
         Args:
-            sock (Socket): Connected socket to target
+            target (session.target): Target we are sending data to
         """
 
-        # default to doing nothing.
-        pass
+        if len(self._pre_send_methods) > 0:
+            try:
+                for f in self._pre_send_methods:
+                    self._fuzz_data_logger.open_test_step('Pre_Send callback: "{0}"'.format(f.__name__))
+                    f(target=target, fuzz_data_logger=self._fuzz_data_logger, session=self, sock=target)
+            except Exception:
+                self._fuzz_data_logger.log_error(constants.ERR_CALLBACK_FUNC.format(func_name="pre_send")
+                                                 + traceback.format_exc())
 
-    def restart_target(self, target):
+    def _restart_target(self, target):
         """
         Restart the fuzz target. If a VMControl is available revert the snapshot, if a process monitor is available
-        restart the target process. Otherwise, do nothing.
+        restart the target process. If custom restart methods are registered, execute them. Otherwise, do nothing.
 
         Args:
             target (session.target): Target we are restarting
 
-        @raise sex.BoofuzzRestartFailedError if restart fails.
+        @raise exception.BoofuzzRestartFailedError if restart fails.
         """
 
-        self._fuzz_data_logger.open_test_step("restarting target")
+        self._fuzz_data_logger.open_test_step("Restarting target")
         if len(self.on_failure) > 0:
             for f in self.on_failure:
-                self._fuzz_data_logger.open_test_step("calling registered on_failure method")
+                self._fuzz_data_logger.open_test_step("Calling registered on_failure method")
                 f(logger=self._fuzz_data_logger)
         # vm restarting is the preferred method so try that before procmon.
         elif target.vmcontrol:
-            self._fuzz_data_logger.log_info("restarting target virtual machine")
+            self._fuzz_data_logger.log_info("Restarting target virtual machine")
             target.vmcontrol.restart_target()
 
         # if we have a connected process monitor, restart the target process.
         elif target.procmon:
-            self._fuzz_data_logger.log_info("restarting target process")
+            self._fuzz_data_logger.log_info("Restarting target process")
 
             if not target.procmon.restart_target():
                 raise exception.BoofuzzRestartFailedError()
 
-            self._fuzz_data_logger.log_info("giving the process 3 seconds to settle in ")
+            self._fuzz_data_logger.log_info("Giving the process 3 seconds to settle in")
             time.sleep(3)
+
+        # if we have custom restart methods, execute them
+        elif len(self._restart_methods) > 0:
+            try:
+                for f in self._restart_methods:
+                    self._fuzz_data_logger.open_test_step('Target restart callback: "{0}"'.format(f.__name__))
+                    f(target=target, fuzz_data_logger=self._fuzz_data_logger, session=self, sock=target)
+            except exception.BoofuzzRestartFailedError:
+                raise
+            except Exception:
+                self._fuzz_data_logger.log_error(constants.ERR_CALLBACK_FUNC.format(func_name="restart_target")
+                                                 + traceback.format_exc())
+            finally:
+                self._fuzz_data_logger.open_test_step('Cleaning up connections from callbacks')
+                target.close()
+                if self._reuse_target_connection:
+                    self._fuzz_data_logger.open_test_step('Reopening target connection')
+                    target.open()
 
         # otherwise all we can do is wait a while for the target to recover on its own.
         else:
             self._fuzz_data_logger.log_info(
-                "no reset handler available... sleeping for %d seconds" % self.restart_sleep_time
+                "No reset handler available... sleeping for {} seconds".format(self.restart_sleep_time)
             )
             time.sleep(self.restart_sleep_time)
 
@@ -983,7 +1055,7 @@ class Session(pgraph.Graph):
                 self._fuzz_data_logger.log_fail(msg)
         try:  # recv
             if self._receive_data_after_each_request:
-                self.last_recv = self.targets[0].recv(10000)  # TODO: Remove magic number (10000)
+                self.last_recv = self.targets[0].recv()
 
                 if self._check_data_received_each_request:
                     self._fuzz_data_logger.log_check("Verify some data was received from the target.")
@@ -1037,7 +1109,7 @@ class Session(pgraph.Graph):
 
         try:  # recv
             if self._receive_data_after_fuzz:
-                self.last_recv = self.targets[0].recv(10000)  # TODO: Remove magic number (10000)
+                self.last_recv = self.targets[0].recv()
         except exception.BoofuzzTargetConnectionReset:
             if self._check_data_received_each_request:
                 self._fuzz_data_logger.log_fail(constants.ERR_CONN_RESET)
@@ -1055,7 +1127,17 @@ class Session(pgraph.Graph):
     def build_webapp_thread(self, port=constants.DEFAULT_WEB_UI_PORT):
         app.session = self
         http_server = HTTPServer(WSGIContainer(app))
-        http_server.listen(port)
+        while True:
+            try:
+                http_server.listen(port)
+            except socket.error as exc:
+                # Only handle "Address already in use"
+                if exc.errno != errno.EADDRINUSE:
+                    raise
+                port += 1
+            else:
+                self._fuzz_data_logger.log_info("Web interface can be found at http://localhost:%d"%port)
+                break
         flask_thread = threading.Thread(target=IOLoop.instance().start)
         flask_thread.daemon = True
         return flask_thread
@@ -1243,12 +1325,13 @@ class Session(pgraph.Graph):
                 target.netmon.pre_send(self.total_mutant_index)
 
             try:
-                target.open()
+                if not self._reuse_target_connection:
+                    target.open()
             except exception.BoofuzzTargetConnectionFailedError:
                 self._fuzz_data_logger.log_error(constants.ERR_CONN_FAILED_TERMINAL)
                 raise
 
-            self.pre_send(target)
+            self._pre_send(target)
 
             for e in path[:-1]:
                 node = self.nodes[e.dst]
@@ -1260,14 +1343,16 @@ class Session(pgraph.Graph):
 
             self._fuzz_data_logger.open_test_step("Node Under Test '{0}'".format(self.fuzz_node.name))
             self.transmit_normal(target, self.fuzz_node, path[-1], callback_data=callback_data)
-            target.close()
 
             self._post_send(target)
             self._check_procmon_failures(target)
+            if not self._reuse_target_connection:
+                target.close()
 
-            self._fuzz_data_logger.open_test_step("Sleep between tests.")
-            self._fuzz_data_logger.log_info("sleeping for %f seconds" % self.sleep_time)
-            time.sleep(self.sleep_time)
+            if self.sleep_time > 0:
+                self._fuzz_data_logger.open_test_step("Sleep between tests.")
+                self._fuzz_data_logger.log_info("sleeping for %f seconds" % self.sleep_time)
+                time.sleep(self.sleep_time)
         finally:
             if self._process_failures(target=target):
                 print("FAIL: {0}".format(test_case_name))
@@ -1311,12 +1396,13 @@ class Session(pgraph.Graph):
 
         try:
             try:
-                target.open()
+                if not self._reuse_target_connection:
+                    target.open()
             except exception.BoofuzzTargetConnectionFailedError:
                 self._fuzz_data_logger.log_error(constants.ERR_CONN_FAILED_TERMINAL)
                 raise
 
-            self.pre_send(target)
+            self._pre_send(target)
 
             for e in path[:-1]:
                 node = self.nodes[e.dst]
@@ -1327,18 +1413,21 @@ class Session(pgraph.Graph):
             callback_data = self._callback_current_node(node=self.fuzz_node, edge=path[-1])
             self._fuzz_data_logger.open_test_step("Fuzzing Node '{0}'".format(self.fuzz_node.name))
             self.transmit_fuzz(target, self.fuzz_node, path[-1], callback_data=callback_data)
-            target.close()
 
             if not self._check_for_passively_detected_failures(target=target):
                 self._post_send(target)
                 self._check_procmon_failures(target=target)
+            if not self._reuse_target_connection:
+                target.close()
 
-            self._fuzz_data_logger.open_test_step("Sleep between tests.")
-            self._fuzz_data_logger.log_info("sleeping for %f seconds" % self.sleep_time)
-            time.sleep(self.sleep_time)
+            if self.sleep_time > 0:
+                self._fuzz_data_logger.open_test_step("Sleep between tests.")
+                self._fuzz_data_logger.log_info("sleeping for %f seconds" % self.sleep_time)
+                time.sleep(self.sleep_time)
         finally:
             self._process_failures(target=target)
             self._stop_netmon(target=target)
+            self._fuzz_data_logger.close_test_case()
             self.export_file()
 
     def _test_case_name_feature_check(self, path):
@@ -1354,13 +1443,9 @@ class Session(pgraph.Graph):
         return "{0}.{1}.{2}".format(message_path, primitive_under_test, self.fuzz_node.mutant_index)
 
     def _post_send(self, target):
-        try:
-            deprecated_callbacks = [self.post_send]
-        except AttributeError:
-            deprecated_callbacks = []
-        if len(self._post_test_case_methods) + len(deprecated_callbacks) > 0:
+        if len(self._post_test_case_methods) > 0:
             try:
-                for f in itertools.chain(self._post_test_case_methods, deprecated_callbacks):
+                for f in self._post_test_case_methods:
                     self._fuzz_data_logger.open_test_step('Post- test case callback: "{0}"'.format(f.__name__))
                     f(target=target, fuzz_data_logger=self._fuzz_data_logger, session=self, sock=target)
             except exception.BoofuzzTargetConnectionReset:
@@ -1368,15 +1453,13 @@ class Session(pgraph.Graph):
             except exception.BoofuzzTargetConnectionAborted as e:
                 self._fuzz_data_logger.log_info(constants.ERR_CONN_ABORTED.format(socket_errno=e.socket_errno,
                                                                                   socket_errmsg=e.socket_errmsg))
-                pass
             except exception.BoofuzzTargetConnectionFailedError:
                 self._fuzz_data_logger.log_fail(constants.ERR_CONN_FAILED)
             except Exception:
-                self._fuzz_data_logger.log_fail(
-                    "Custom post_send method raised uncaught Exception." + traceback.format_exc())
+                self._fuzz_data_logger.log_error(constants.ERR_CALLBACK_FUNC.format(func_name="post_send")
+                                                 + traceback.format_exc())
             finally:
                 self._fuzz_data_logger.open_test_step('Cleaning up connections from callbacks')
-                target.close()
 
     def _reset_fuzz_state(self):
         """