+++ unit_tests/test_session_failure_handling.py
@@ -163,8 +163,9 @@ class TestNoResponseFailure(unittest.TestCase):
             ),
             fuzz_loggers=[],  # log to nothing
             check_data_received_each_request=True,
+            keep_web_open=False,
         )
-        session.restart_target = self._mock_restart_target()
+        session._restart_target = self._mock_restart_target()
 
         s_initialize("test-msg-a")
         s_string("test-str-value")